Ran brute 2**20 search. Hm...
So the 2**20 search returned 8 valid assignments.
But these assignments have a lot of conflicts in the independent set
graph.

This must be because, the IS graph makes it impossible to select two
literals from the same clause.

The valid subsets search returned a solution of 11 variables.

Is this reconcilable?... Needs more thought.